+++ includes/gateways/class-wpsms-gateway-mitto.php
@@ -12,7 +12,7 @@ class mitto extends \WP_SMS\Gateway
     public $unit;
     public $flash = "disable";
     public $isflash = false;
-    public $documentUrl = 'https://wsms.io/docs/gateway-configuration';
+    public $documentUrl = 'https://wsms.io/docs/mitto-gateway-setup';
     public $gatewayFields = [
         'from'    => [
             'id'           => 'gateway_sender_id',